ファイル構成の変更とサインインとサインアップ画面のの表示
